Corrections suite au passage sous 64bits
git-svn-id: svn://team-ard.com/ARP/trunk@658 c9b90f90-cbc1-11df-9d65-af01d29d43da

+++ arp_hml/test/orocos/TestJoystick.cpp
@@ -27,8 +27,8 @@ TaskContext* tcJoystick;
 TestJoystick::TestJoystick(const std::string& name):
 	HmlTaskContext(name)
 {
-	int attrTcTest = (int) tcTest;
-	int attrTcJoystick = (int) tcJoystick;
+	long attrTcTest = (long) tcTest;
+	long attrTcJoystick = (long) tcJoystick;
 	addAttribute("tcTest", attrTcTest);
 	addAttribute("tcJoystick", attrTcJoystick);
